+++ src/index.ts
@@ -30,9 +30,10 @@ export { default as UxRUL } from "./components/RUL";
 export { default as UxCollapse } from "./components/Collapse";
 export { default as UxPopover } from "./components/Popover";
 export { default as UxTextArea } from "./components/TextArea";
-export { default as UxEnergyBarWidget } from "./components/EnergyBarWidget";
 export { default as UxFileUpload } from "./components/FileUpload";
 export { default as UxDatePicker } from "./components/DatePicker";
+export { default as UxEnergyBarWidget } from "./components/EnergyBarWidget";
+export { default as UxEnergyBarGraph } from "./components/EnergyBarGraph";
 
 //Props
 export { PMActionProps, ContentProps } from "./types/pmcalendar";